requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
mysqlclient==2.2.6
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8